CacheStrategy: refactor base class into `Structures`
