+++ Sources/Enumerations/CacheStrategy.js
@@ -1,7 +1,5 @@
 
-const { Collection } = require("qulity");
-
-const Generics = require("../Generics");
+const BaseCacheStrategy = require("../Structures/BaseCacheStrategy");
 
 class CacheStrategy {
 
@@ -44,32 +42,10 @@ class CacheStrategy {
     /**
      * A caching strategy with no sweeping or automatic eviction logic
      * implemented.
-     * @returns {CacheStrategy}
+     * @returns {BaseCacheStrategy}
      */
     static unlimited() {
-        return new CacheStrategy();
-    }
-
-    // Abstract default implementation
-
-    /**
-     * In-memory cached rows.
-     * @name CacheStrategy#memory
-     * @type {Collection<String, DataModel>}
-     * @private
-     */
-    memory = new Collection();
-
-    /**
-     * Inserts or patches something in the Connection's internal cache.
-     * @param {String} keyContext As address to memory map this data model to.
-     * @param {DataModel} document The value to set in the cache, as a parsed memory model.
-     * @abstract
-     */
-    patch(keyContext, document) {
-        const documentClone = Generics.clone(document);
-        documentClone._timestamp = Date.now();
-        this.memory.set(keyContext, documentClone);
+        return new BaseCacheStrategy();
     }
 }
